backend/requirements.txt
# FastAPI & Server
fastapi>=0.104.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# LangChain 1.x Core (Dec 2025 stable)
langchain>=1.1.0
langchain-core>=1.1.0
langchain-community>=0.4.0
langchain-text-splitters>=1.0.0
langgraph>=1.0.0

# Embeddings
langchain-huggingface>=1.1.0
sentence-transformers>=3.0.0
torch>=2.3.0

# Vector Store
langchain-chroma>=1.0.0
chromadb>=1.3.0

# LLM - Google Gemini
langchain-google-genai>=3.2.0
google-generativeai>=0.8.0,<0.9.0   # pinned to avoid future breakage

# PDF Processing
pypdf>=4.0.0

# Utilities
tqdm>=4.66.0
numpy>=1.26.0
pydantic>=2.7.0
httpx>=0.27.0

# Optional: silence Hugging Face symlink warning on Windows
# (remove if you enable Developer Mode or run as admin)
huggingface-hub>=0.23.0